ENH: Move BIDS IO to separate module
