+++ BIDSIO/Testing/Python/CMakeLists.txt
@@ -1,0 +1,2 @@
+
+#slicer_add_python_unittest(SCRIPT ${MODULE_NAME}ModuleTest.py)
